requirements.txt
flask
pillow
torch
yolov5